lots of changes
data_processing deals w reading in list of candidate words and words used in the past.

game logic applies the logic from the hints to narrow down list of candidate words.

website_handling deals with handling the javascript from the web page i.e. reading off the tiles and writing to them

wordle_doc is the original file that i am reading in past words and candidate wrods from

wordle_doc2 is a copy of wordle_doc that I am writing to because I don't want to risk messing up the original file